program to analyse palindrome and anagram
A palindrome is a word, phrase, number, or other sequence of characters which reads the same
backward or forward, an example of this is the name "Anna" or the sentence "Do not bob to nod". An
anagram is a type of word play which is the result of rearranging the letters of a word or phrase to
produce a new word or phrase an example of this is the words "silent" and "listen" are composed of
the same letters.
This task is to write a method which accepts a string as a parameter and returns a boolean. The
method will return true if the string is a palindrome or an anagram of a palindrome otherwise it will
return false.